requirements.txt
fastapi
uvicorn[standard]
tensorflow
pillow
numpy
opencv-python-headless
python-multipart
